+++ addOns/authhelper/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added support for Client Script Authentication when used in conjunction with the Ajax Spider add-on.
 - Add support for custom authentication steps in Browser Based Authentication.
 
+### Fixed
+- Reset always the state of the demo mode in the Authentication Tester dialogue.
+
 ## [0.18.0] - 2025-01-27
 ### Changed
 - Ignore non-displayed fields when selecting the user name and password.